[aws] ec2 instance creation and initialization
Summary:

This diff adds a set of scripts for single-invocation instance creation
and initialization.

Test Plan:

From my local host, with an appropriate user SSH config installed,
confirm that the setup scripts execute as expected, that the instance
appears as expected in the AWS console, and that project repositories
have been cloned on the remote host.
```bash
 ~/Developer/projects/infra/python/ktd/aws/ec2/create_dev_instance.py
 ssh dev-main 'ls -la $HOME/projects'
```