[aws] idle threshold < 0.5% + detailed monitoring
Summary:
This commit reduces the idle threshold to 0.5% CPU utilization after
some premature devserver stoppages. It also adds detailed monitoring to
the AWS CloudFormation template for devservers; it's cheap and will
likely reduce the chance of sampling error for the inactive alarm. and
given that the price for alarms with < 1m period when detailed
monitoring is enabled is higher, we also increase the period to 1m.

Test Plan:
Confirm CPU usage during a low-activity period - file
editing, etc. - is above 0.5% and when idle, below 0.5%. Confirm that
detailed monitoring is enabled in the AWS console and that alarm plots
show the higher temporal resolution data.