[ray] save repo state in ckpt, update group names
Summary:

This commit adds the ability to save the state of the git repository in
checkpoints for reproducibility. With this change, it's no longer useful
to include that repo state in the logs, which was done via the config,
so we remove the full state from the config and instead include a
summary in the train_loop_config itself, which can be used to filter and
group trials in wandb. With this, we can just use date/time as the group
name.

Test Plan:
In a tuning run, confirm that
- the repo state summary appears in the driver logs
- the trial path in S3 is as expected
- the group name in wandb is as expected
- `Checkpoint.from_directory(dir).get_metadata()` contains `repo_state`