[sys] cleanup venv prompt and activation
Summary:
Previously we were activating the default venv for each login shell, and
reactivating for any interactive shell. This had unintended consequences
on the PATH and on PS1. Here we avoid activating a venv that's already
active and manage the venv component of PS1 ourselves. We also need
to ensure that, at the conclusion of sh.local, we activate the default
venv; so we de-activate before activating the default venv in sh.local
to ensure the path is properly updated.

Test Plan:

Confirm that:
- the venv prompt prefix appears as expected
- /opt/homebrew/bin appears in the PATH in the expected order (e.g.,
before /usr/local/bin) after opening login and non-login interactive
shells
- elements of PATH are ordered as expected in a vscode terminal that
activates the default venv beforehand
```