[util] hydra helpers
Summary:
This commit adds a few utilities for Hydra, to be used in the following
commits for moving experiments over to Hydra. The most relevant is a
custom `instantiate()` with support for deferring resolution and
instantiation for select config nodes, useful for hyperparameter tuning
runs where we want to defer parameter instantiation until trial-specific
values are known, and full instantiation to primitive containers with
arbitrary types (not just `dataclass`).

Test Plan:
See tests added to, and test plan for, the last commit in this PR.